Get default lattice from system in a separate function, separate phase and phase list files
