+++ orix/io/plugins/ctf.py
@@ -28,8 +28,8 @@ from typing import Any, Literal
 from diffpy.structure import Lattice, Structure
 import numpy as np
 
+from orix.crystal_map._phase_list import PhaseList
 from orix.crystal_map.crystal_map import CrystalMap, _data_slices_from_coordinates
-from orix.crystal_map.phase_list import PhaseList
 from orix.quaternion.rotation import Rotation
 
 __all__ = ["file_reader"]
